Binary Search Tree (BST) with Traversals: Recursive Node Insertion
This C++ program implements a Binary Search Tree (BST) with recursive insertion of a sequence of nodes entered by the user. After the tree is built, the program displays the results using three traversal methods:

Preorder

Inorder

Postorder


A simplified vertical visualization of the tree is also included to observe how nodes are hierarchically constructed.